Fixes to make build work on a virgin install (see below for details)
Disclaimer: I'm a total noob to nodeish things and I just found your great
project.  Alas, I had some problems building from the current version in
master (sha e220ab):

* yarn install failed due to conflicting versions for a few library
dependencies.  So this changes the relevant package.json files to resolve
those conflicts (by picking the highest version # used)
* yarn run dev failed due to error messages re conflicting yarn.lock files.
I didn't fully understand but stackoverflow said that when using workspaces
in yarn it is important to delete the yarn.lock files from the subdirectories
(and not put them in source control) - because it will keep the same info in the
root yarn.lock.  I'm not sure if this is correct but doing so fixed the build
for me.  What do you think?
* Fix the documentation build by updating docusaurus and resolving incompatible
version numbers.
* The instructions for how to build the protobuf glue was slightly stale
(based on looking at package.json)